+++ NEWS.md
@@ -19,7 +19,9 @@
 * Added `docvars()` and `docvars<-()` methods for `tokens_xptr` objects,
   allowing document variables to be read and assigned without recompiling the
   tokens. (#2497)
-  
+
+* Add the `drop0()` method for dfm and fcm (#2460).
+
 # quanteda 4.4
 
 ## Changes and additions